BigInteger id use long type

+++ src/main/java/com/fish/myspringboot/mapper/UserMapper.java
@@ -4,19 +4,18 @@ import com.fish.myspringboot.pojo.User;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Mapper
 public interface UserMapper {
 
-    User getById(@Param("id") BigInteger id);
+    User getById(@Param("id") long id);
 
     int updateUser(User user);
 
     int insertUser(User user);
 
-    int deleteUser(@Param("id") int id);
+    int deleteUser(@Param("id") long id);
 
     List<User> getAllUsers();
 }
